tutorials: rework "input_gui_read_only" to split code into three different files.
