+++ doc/tutorials/my_first_widget/materials/input_gui_read_only/OWClipData.py
@@ -1,0 +1,34 @@
+"""
+OWClipData.py: Code for the orange add-on binding.
+"""
+
+from ewoksorange.gui.owwidgets.threaded import OWEwoksWidgetOneThread
+from ewokstesttuto.tasks.clipdata import ClipDataTask
+from ewokstesttuto.gui.MyWidget import MyWidget
+from ewokscore.missing_data import is_missing_data
+
+
+class OWClipData(
+    OWEwoksWidgetOneThread,
+    ewokstaskclass=ClipDataTask,
+):
+    name = "rescale data"
+    id = "orange.widgets.my_project.ClipDataTask"
+    description = "widget to clip data (numpy array) within a percentile range."
+    want_main_area = True
+    want_control_area = False
+
+    _ewoks_inputs_to_hide_from_orange = ("percentiles",)
+
+    def __init__(self, parent=None):
+        super().__init__(parent)
+
+        self._myWidget = MyWidget(self)
+        self.mainArea.layout().addWidget(self._myWidget)
+
+    def handleNewSignals(self):
+        percentiles = self.get_task_input_value("percentiles")
+        if not is_missing_data(percentiles):
+            self._myWidget.setPercentiles(percentiles)
+
+        return super().handleNewSignals()
